shittiest skinning ever. need to integrate it with paging system + write up some kind of test animator to make sure it works properly

+++ src/asset/asset_model.h
@@ -17,7 +17,7 @@ struct AST_ModelVertex
 typedef struct AST_ModelSkinVertex AST_ModelSkinVertex;
 struct AST_ModelSkinVertex
 {
-	u16 joints[4];
+	u32 joints[4];
 	f32 weights[4];
 };
